requirements.txt
# requirements.txt
python-telegram-bot==20.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.1.2
transformers==4.36.0
Pillow==10.1.0
openai==1.3.9
requests==2.31.0
pytest==7.4.3
black==23.12.0
ruff==0.1.11
python-dotenv==1.0.0
